pcrs/requirements.txt
Django==1.11
django-appconf==1.0.2
django-common-helpers==0.9.1
django-compressor==2.1.1
django-crispy-forms==1.6.1
django-cron==0.5.0
djangorestframework==3.6.2
olefile==0.44
pep8==1.7.0
Pillow==4.2.1
psycopg2==2.7.1
pycparser==2.17
pyparsing==2.2.0
pytz==2017.2
rapt==0.0.2
rcssmin==1.0.6
reportlab==3.4.0
requests==2.13.0
rjsmin==1.0.12
six==1.10.0
python_ta==1.2.0
